LayoutTests/accessibility/mac/attributed-string/attributed-string-for-range-expected.txt

word mispelled word
This test ensures that attributed string for text marker range works and does include misspelled attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS text is expected
PASS successfullyParsed is true

TEST COMPLETE
